pyproject.toml
[tool.poetry]
name = "humanornot"
version = "0.1.0"
description = ""
authors = ["sbergsmann <29523118+sbergsmann@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
openai = "^1.17.0"
pydantic = "^2.6.4"
instructor = "^1.0.3"
flet = "^0.22.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
